Moving sec_mobile_main to new file, shrinking text in tables
